src/pyproject.toml
[tool.poetry]
name = "portfolio-blog"
version = "0.1.0"
description = ""
authors = ["Aykhan <aykhan.shahs0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
psycopg2 = "^2.9.7"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.3.0"}
python-multipart = "^0.0.6"
python-slugify = "^8.0.1"
pillow = "^10.0.0"
aiofiles = "^23.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
